requirements.txt
python-dotenv
langchain_core
langchain_groq
langchain
langchain_community
pymysql
streamlit
